+++ CHANGES.md
@@ -29,6 +29,8 @@
 
 <!-- Changes to blackd -->
 
+- `blackd` now supports preview style via `X-Preview` header (#3217)
+
 ### Configuration
 
 <!-- Changes to how Black can be configured -->